Define create app mutation
